refactor(FilterChannels): split the component in smaller components
