--- webapp/src/AdminSettings/FilterChannels.scss
@@ -1,161 +1,0 @@
-.FilterChannels {
-  .EnableFilter {
-    display: flex;
-
-    label {
-      display: inline-flex;
-      opacity: 1;
-      margin-top: 8px;
-      margin-right: 24px;
-      width: fit-content;
-      flex-direction: row;
-      -webkit-box-align: center;
-      align-items: center;
-      margin-bottom: 0px;
-      cursor: pointer;
-      font-size: 14px;
-      font-weight: 400;
-      gap: 8px;
-      line-height: 20px;
-    }
-
-    input {
-      display: grid;
-      width: 1.6rem;
-      height: 1.6rem;
-      border: 1px solid rgba(var(--center-channel-color-rgb), 0.24);
-      border-radius: 50%;
-      margin: 0px;
-      appearance: none;
-      background-color: white;
-      color: rgba(var(--center-channel-color-rgb), 0.24);
-      cursor: pointer;
-      font: inherit;
-      place-content: center;
-
-      &:checked {
-        border-color: var(--denim-button-bg);
-      }
-
-      &::before {
-        width: 8px;
-        height: 8px;
-        border-radius: 50%;
-        background: var(--denim-button-bg);
-        content: "";
-        transform: scale(0);
-        transform-origin: center center;
-        transition: transform 200ms ease-in-out;
-      }
-
-      &:checked::before {
-        transform: scale(1);
-      }
-    }
-
-    span {
-      cursor: pointer;
-      font-size: 14px;
-      font-weight: 400;
-      line-height: 20px;
-    }
-  }
-
-  .ChannelSelector {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 8px;
-    margin-top: 32px;
-    height: 20em;
-
-    .ChannelSelectorColumn {
-      display: flex;
-      flex: 1;
-      flex-direction: column;
-      background: #111;
-      border-radius: 4px;
-      max-height: 100%;
-
-      input {
-        cursor: pointer;
-      }
-
-      .ChannelSelectorHeader {
-        display: flex;
-        padding: var(--list-small-spacing);;
-        align-items: center;
-        background-color: rgba(255,255,255,.15);
-
-        h3 {
-          font-size: 1em;
-          font-weight: bold;
-          margin: 0;
-          padding: 0;
-        }
-
-        .Subtitle {
-          font-weight: normal;
-          opacity: 0.7;
-        }
-
-        input:indeterminate + span {
-          display: flex;
-          justify-content: center;
-          align-items: center;
-        
-          &::after {
-            content: "";
-            width: 70%;
-            border-bottom: 2px solid #1e3a54;
-          }
-        }
-      }
-
-      .ChannelSelectorList {
-        padding: 0;
-        overflow-y: auto;
-
-        .ChannelSelectorListLink {
-          border-radius: 0;
-          
-          > div {
-            padding: 0;
-          }
-
-          &:hover {
-            background-color: rgba(255,255,255,.15);
-          }
-
-          span {
-            line-height: 0;
-            width: 100%;
-          }
-
-          label {
-            align-items: center;
-            margin: 0;
-            padding: var(--list-small-spacing);
-            width: 100%;
-          }
-        }
-      }
-    }
-
-    .ChannelSelectorButtons {
-      display: flex;
-      flex-direction: column;
-      gap: 1em;
-      justify-content: center;
-      max-width: 3em;
-
-      button {
-        background-color: rgb(17, 17, 17);
-        box-shadow: none !important;
-
-        &:hover {
-          background-color: rgba(17, 17, 17, .7);
-        }
-      }
-    }
-  }
-}
