+++ webapp/src/AdminSettings/FilterChannels/components/ToggleEnable/ToggleEnable.scss
@@ -1,0 +1,61 @@
+.ToggleEnable {
+  display: flex;
+
+  label {
+    display: inline-flex;
+    opacity: 1;
+    margin-top: 8px;
+    margin-right: 24px;
+    width: fit-content;
+    flex-direction: row;
+    -webkit-box-align: center;
+    align-items: center;
+    margin-bottom: 0px;
+    cursor: pointer;
+    font-size: 14px;
+    font-weight: 400;
+    gap: 8px;
+    line-height: 20px;
+  }
+
+  input {
+    display: grid;
+    width: 1.6rem;
+    height: 1.6rem;
+    border: 1px solid rgba(var(--center-channel-color-rgb), 0.24);
+    border-radius: 50%;
+    margin: 0px;
+    appearance: none;
+    background-color: white;
+    color: rgba(var(--center-channel-color-rgb), 0.24);
+    cursor: pointer;
+    font: inherit;
+    place-content: center;
+
+    &:checked {
+      border-color: var(--denim-button-bg);
+    }
+
+    &::before {
+      width: 8px;
+      height: 8px;
+      border-radius: 50%;
+      background: var(--denim-button-bg);
+      content: "";
+      transform: scale(0);
+      transform-origin: center center;
+      transition: transform 200ms ease-in-out;
+    }
+
+    &:checked::before {
+      transform: scale(1);
+    }
+  }
+
+  span {
+    cursor: pointer;
+    font-size: 14px;
+    font-weight: 400;
+    line-height: 20px;
+  }
+}
