+++ webapp/src/types/ChannelWithChecked.ts
@@ -1,0 +1,3 @@
+import type { Channel } from 'mattermost-redux/types/channels'
+
+export type ChannelWithChecked = Channel & { checked: boolean }
